Template and health also moved
